+++ src/impls/string.rs
@@ -33,6 +33,16 @@ pub fn impl_string(info: &Info) -> proc_macro2::TokenStream {
         });
     }
 
+    if info.kind == Kind::BoxStr {
+        tokens.extend(quote! {
+            impl ::core::convert::From<String> for #name {
+                fn from(s: String) -> Self {
+                    Self(s.into_boxed_str())
+                }
+            }
+        });
+    }
+
     tokens
 }